requirements.txt
black==21.9b0
eth-brownie>=1.11.0,<2.0.0
dotmap==1.3.24
python-dotenv==0.16.0
tabulate==0.8.9
rich==10.7.0
click==8.0.1
platformdirs==2.3.0
regex==2021.8.28